+++ src/utils.cpp
@@ -273,7 +273,7 @@ SATargetType getScriptActionTargetType(uint8_t ID)
 	}
 }
 
-size_t getStringListMaxWidth(QStringList list, QFont font)
+int32_t getStringListMaxWidth(QStringList list, QFont font)
 {
 	if (!list.empty()) {
 		std::vector <size_t> widths;
@@ -281,12 +281,12 @@ size_t getStringListMaxWidth(QStringList list, QFont font)
 		for(QStringList::Iterator listIT = list.begin(); listIT != list.end(); ++listIT) {
 			widths.push_back(metr.width(*listIT));
 		}
-		return *max_element(widths.begin(), widths.end());
+		return int32_t(*max_element(widths.begin(), widths.end()));
 	}
 	return 0;
 }
 
-size_t getComboBoxMaxWidth(QComboBox *box)
+int32_t getComboBoxMaxWidth(QComboBox *box)
 {
 	box->update();
 	if (box != NULL && box->count() > 0) {
@@ -295,7 +295,7 @@ size_t getComboBoxMaxWidth(QComboBox *box)
 		for(int i = 0; i < box->count(); ++i) {
 			widths.push_back(metr.width(box->itemText(i)));
 		}
-		return *max_element(widths.begin(), widths.end());
+		return int32_t(*max_element(widths.begin(), widths.end()));
 	}
 	return 0;
 }
